task-3/requirements.txt
streamlit
pybase64
langchain
python-dotenv
pillow
Python-IO
langchain
langchain_core
langchain_google_genai